updates and changes

+++ src/services/question/RemoveQuestionService.ts
@@ -1,6 +1,9 @@
 import questionRepository from "../../repositories/questionRepository";
 import { BadRequestError } from "../../utils/ApiErrors";
+import Question from '../../entities/Question';
 import paramsConfig from "../../params/paramsConfig";
+import QuestionGroupMapping from '../../entities/QuestionGroupMapping';
+import questionGroupMappingRepository from '../../repositories/questionGroupMappingRepository';
 
 type QuestionRequest =
 {
@@ -9,16 +12,20 @@ type QuestionRequest =
 
 class RemoveQuestionService
 {
-	public async execute({ id }: QuestionRequest)
+	public async execute({ id }: QuestionRequest): Promise<string>
 	{
-		const question = await questionRepository.findOneBy({ id: Number(id) });
+		if(!paramsConfig.params.allowRemoveQuestions) {
+			throw new BadRequestError('not-allowed-delete-question');
+		}
+
+		const question: Question | null = await questionRepository.findOneBy({ id: Number(id) });
 		if(!question) {
 			throw new BadRequestError('no-question');
 		}
 
-		if(!paramsConfig.params.allowRemoveQuestions) {
-			throw new BadRequestError('not-allowed-delete-question');
-		}
+		const checkQuestion: QuestionGroupMapping | null = await questionGroupMappingRepository.findOne({ where: { question: { id: Number(id)} } });
+		if(checkQuestion)
+			throw new BadRequestError('this-question-is-already-being-used');
 
 		await questionRepository.remove(question);
 		return 'question-removed';
